added AvatarDisplay and menu in DashboardHeader
