--- client/src/components/AuthMenu/AuthMenu.tsx
@@ -1,79 +1,0 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-import { useState, MouseEvent } from 'react';
-import IconButton from '@material-ui/core/IconButton';
-import Menu from '@material-ui/core/Menu';
-import MenuItem from '@material-ui/core/MenuItem';
-import MoreHorizIcon from '@material-ui/icons/MoreHoriz';
-import { useAuth } from '../../context/useAuthContext';
-
-declare global {
-  interface Window {
-    cloudinary: any;
-  }
-}
-
-const AuthMenu = (): JSX.Element => {
-  const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
-  const open = Boolean(anchorEl);
-  const { logout, loggedInUser, updateUser } = useAuth();
-
-  const handleClick = (event: MouseEvent<HTMLElement>) => {
-    setAnchorEl(event.currentTarget);
-  };
-
-  const handleClose = () => {
-    setAnchorEl(null);
-  };
-
-  const handleLogout = () => {
-    handleClose();
-    logout();
-  };
-
-  const openWidget = () => {
-    // create the Cloudinary upload widget
-    window.cloudinary
-      .createUploadWidget(
-        {
-          cloudName: process.env.REACT_APP_CLOUDINARY_CLOUD_NAME,
-          uploadPreset: process.env.REACT_APP_CLOUDINARY_UPLOAD_PRESET,
-        },
-        (error: any, { event, info }: any) => {
-          if (event === 'success' && loggedInUser != null) {
-            updateUser({
-              ...loggedInUser,
-              profilePicture: info.secure_url,
-            });
-          }
-        },
-      )
-      .open(); // opens upload dialog after widget creation
-  };
-
-  return (
-    <div>
-      <IconButton aria-label="show auth menu" aria-controls="auth-menu" aria-haspopup="true" onClick={handleClick}>
-        <MoreHorizIcon />
-      </IconButton>
-      <Menu
-        id="auth-menu"
-        anchorEl={anchorEl}
-        keepMounted
-        open={open}
-        onClose={handleClose}
-        anchorOrigin={{
-          vertical: 'bottom',
-          horizontal: 'left',
-        }}
-        getContentAnchorEl={null}
-      >
-        <MenuItem className="widget-btn" onClick={openWidget}>
-          Change profile picture
-        </MenuItem>
-        <MenuItem onClick={handleLogout}>Logout</MenuItem>
-      </Menu>
-    </div>
-  );
-};
-
-export default AuthMenu;
